backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
setuptools>=65.0.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Cache & Task Queue
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0

# File Storage
boto3==1.34.34
django-storages==1.14.2
minio==7.2.0

# Configuration
python-decouple==3.8
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.61.1
firebase-admin>=6.0.0  # Firebase 토큰 검증용

# AI Services
requests==2.32.3

# Payments

# WSGI Server
gunicorn==21.2.0

# Development
django-debug-toolbar==4.2.0
